requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
grpcio==1.67.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.8
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.1
propcache==0.2.0
protobuf==5.28.3
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.5.0
torchmetrics==1.5.1
tqdm==4.66.5
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.16.0
